+++ src/main/java/got/client/effect/GOTEntityRiverWaterFX.java
@@ -27,6 +27,6 @@ public class GOTEntityRiverWaterFX extends EntitySpellParticleFX {
 	@Override
 	public void onUpdate() {
 		super.onUpdate();
-		particleAlpha = 0.5f + 0.5f * ((float) particleAge / (float) particleMaxAge);
+		particleAlpha = 0.5f + 0.5f * ((float) particleAge / particleMaxAge);
 	}
 }
