+++ src/main/java/got/common/command/GOTCommandTime.java
@@ -40,7 +40,7 @@ public class GOTCommandTime extends CommandBase {
 		if (args.length >= 2) {
 			if ("set".equals(args[0])) {
 				long time;
-				time = "day".equals(args[1]) ? Math.round(GOTTime.DAY_LENGTH * 0.03) : "night".equals(args[1]) ? Math.round(GOTTime.DAY_LENGTH * 0.6) : (long) CommandBase.parseIntWithMin(sender, args[1], 0);
+				time = "day".equals(args[1]) ? Math.round(GOTTime.DAY_LENGTH * 0.03) : "night".equals(args[1]) ? Math.round(GOTTime.DAY_LENGTH * 0.6) : CommandBase.parseIntWithMin(sender, args[1], 0);
 				GOTTime.setWorldTime(time);
 				CommandBase.func_152373_a(sender, this, "got.command.time.set", time);
 				return;
