+++ src/main/java/got/common/tileentity/GOTTileEntityUnsmeltery.java
@@ -330,7 +330,7 @@ public class GOTTileEntityUnsmeltery extends GOTTileEntityAlloyForge {
 		float items = result.stackSize;
 		items *= 0.8f;
 		if (itemstack.isItemStackDamageable()) {
-			items *= (float) (itemstack.getMaxDamage() - itemstack.getItemDamage()) / (float) itemstack.getMaxDamage();
+			items *= (float) (itemstack.getMaxDamage() - itemstack.getItemDamage()) / itemstack.getMaxDamage();
 		}
 		items_int = Math.round(items *= MathHelper.randomFloatClamp(unsmeltingRand, 0.7f, 1.0f));
 		if (items_int <= 0) {
